fix: correct y offset (#1)
Currently the header expects to be appended to an ion-content element with a top of 0.

This change makes it so that instead of using the offsetTop of the scroll element to determine the y transform of the header, the difference between the top of the scroll element and the top of the element the header will be placed in will be used.

This means we can use nested ion-content elements. Technically you're not supposed to use nested ion-content elements, but I've used one as a work-around to allow a nicely transitioning filter component in the events list for OurPeople/op-app#680.